adicion de controller y vistas de Producto e instalacion de javax.validation

+++ ingresos_egresos/src/main/java/servicios/exequiales/ingresosyegresos/ingresos_egresos/Service/IProductoService.java
@@ -6,9 +6,9 @@
 
  public interface IProductoService {
 
-     public Producto findById(int id);
+     public Producto findById(long id);
      public List<Producto> findAll();
      public Producto creatProducto(Producto producto);
      public Producto updateProducto(int id, Producto producto);
-     public void deletProducto(int id);
+     public void deletProducto(long id);
  }
